1. 完成代码生成工具 genrpc.lua.花了将近一天半时间,眼泪都要掉下来! 2. 将allbeans与allhandlers的Map key类型由short改成int

+++ fiber/src/fiber/app/client/Client.java
@@ -26,7 +26,7 @@ public class Client {
 			final int TASK_NUM = Integer.parseInt(args[3]);
 			
 			final IOPoller poller = new IOPoller(1);
-			final Map<Short, BeanHandler<?>> handlerStub = AllHandlers.get();
+			final Map<Integer, BeanHandler<?>> handlerStub = AllHandlers.get();
 
 			ArrayList<ClientManager> clients = new ArrayList<ClientManager>();
 			for(int i = 0 ; i < CONNECT_NUM ; i++) {
